# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Referral API",
        "description": "API for managing referrals between business units",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        },
        {
            "url": "http://mytotalhealth.com.my/referral-api",
            "description": "MyHealth Server (Staging)"
        }
    ],
    "paths": {
...
        "/api/form": {
            "post": {
                "tags": [
                    "Forms"
                ],
                "summary": "Create a new form and its details",
                "operationId": "a45ee9957cd963fad67e9a6ef5dc11fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "business_unit_id",
                                    "label_name",
                                    "is_hidden",
                                    "is_required",
                                    "field_type"
                                ],
                                "properties": {
                                    "business_unit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "label_name": {
                                        "type": "string",
                                        "example": "Physio Interventions"
                                    },
                                    "is_hidden": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "is_required": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "field_name": {
                                        "type": "string",
                                        "example": "physio_interventions"
                                    },
                                    "field_type": {
                                        "type": "string",
                                        "example": "checkbox"
                                    },
                                    "value_fields": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Chest Physiotherapy",
                                            "Pain Management"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Form created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Form created successfully!"
                                        },
                                        "form_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Authorization header is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Form creation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Form creation failed"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/form/show": {
            "get": {
                "tags": [
                    "Forms"
                ],
                "summary": "Get forms by business unit id from JWT token",
                "operationId": "1114638e3b0805ac69504dbeb41ba42e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "business_unit": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "forms": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "form_id": {
                                                        "type": "integer",
                                                        "example": 12
                                                    },
                                                    "label_name": {
                                                        "type": "string",
                                                        "example": "Assessment Form"
                                                    },
                                                    "is_hidden": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "form_details": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "form_detail_id": {
                                                                    "type": "integer",
                                                                    "example": 101
                                                                },
                                                                "field_name": {
                                                                    "type": "string",
                                                                    "example": "Blood Pressure"
                                                                },
                                                                "field_type": {
                                                                    "type": "string",
                                                                    "example": "text"
                                                                },
                                                                "is_required": {
                                                                    "type": "boolean",
                                                                    "example": true
                                                                },
                                                                "field_value": {
                                                                    "oneOf": [
                                                                        {
                                                                            "type": "string",
                                                                            "example": "120/80"
                                                                        },
                                                                        {
                                                                            "type": "array",
                                                                            "items": {
                                                                                "properties": {
                                                                                    "form_detail_id": {
                                                                                        "type": "integer",
                                                                                        "example": 201
                                                                                    },
                                                                                    "field_value": {
                                                                                        "type": "string",
                                                                                        "example": "Option A"
                                                                                    }
                                                                                },
                                                                                "type": "object"
                                                                            }
                                                                        }
                                                                    ]
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Business unit ID not found in token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid or expired token."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "SQLSTATE[42S22]: Column not found: 1054 Unknown column..."
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/form/{form}": {
            "delete": {
                "tags": [
                    "Forms"
                ],
                "summary": "Delete a form of a business unit",
                "description": "Delete a specific form by its ID.",
                "operationId": "be84502d0c043af16325cb48e41bee12",
                "parameters": [
                    {
                        "name": "form",
                        "in": "path",
                        "description": "The ID of the form to be deleted.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Form deleted successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid or expired token."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete form",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to delete form."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message from exception"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token in the format: Bearer {token}"
            }
        }
    },
...
}